Remove proxy and shift proxy tests to test the cache
