+++ test/tests.ts
@@ -62,4 +62,4 @@ import './cloneDeep';
 import './assign';
 import './environment';
 import './ApolloClient';
-import './proxy';
+import './cache';
